- src folder's content moved one top level - renamed some docs - added top-level file to export components
