+++ components/tag-input/tag-input.component.ts
@@ -17,8 +17,9 @@ import {
 } from './tag-input.constants';
 
 import {NG_VALUE_ACCESSOR} from "@angular/common";
-import {Tag} from './tag/tag.component';
+import {Tag} from '../tag/tag.component';
 import {TagInputAccessor} from './tag-input-accessor';
+import {TagInputComponent} from './tag-input.component.d.ts';
 
 const CUSTOM_INPUT_CONTROL_VALUE_ACCESSOR = new Provider(NG_VALUE_ACCESSOR, {
     useExisting: forwardRef(() => TagInput),
@@ -37,7 +38,7 @@ const CUSTOM_INPUT_CONTROL_VALUE_ACCESSOR = new Provider(NG_VALUE_ACCESSOR, {
     template: require('./tag-input.template.html'),
     changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class TagInput extends TagInputAccessor implements tagInput.TagInputComponent {
+export class TagInput extends TagInputAccessor implements TagInputComponent {
     /**
      * @name separatorKeys
      * @desc keyboard keys with which a user can separate items
