+++ components/tag/tag.component.d.ts
@@ -1,0 +1,3 @@
+export interface TagComponent {
+    remove(): void;
+}
