+++ components/tag/tag.component.ts
@@ -1,11 +1,12 @@
 import {Component, Input, Output, EventEmitter} from '@angular/core';
+import {TagComponent} from './tag.component.d.ts';
 
 @Component({
     selector: 'tag',
     styles: [require('./tag.style.scss').toString()],
     template: require('./tag.template.html')
 })
-export class Tag implements tagInput.TagComponent {
+export class Tag implements TagComponent {
     @Input() item: string;
     @Input() readonly: boolean;
     @Input() isSelected: boolean;
