+++ components/tag/tag.template.html
@@ -3,6 +3,6 @@
         {{ item }}
     </span>
     <span class="tag__remove-button" (click)="remove()" *ngIf="!readonly">
-        <img src="./delete.png" />
+        <img src="delete.png" />
     </span>
 </div>
