+++ demo/index.html
@@ -4,7 +4,7 @@
     <meta charset=UTF-8>
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <title>Ng2 Tag Input Demo</title>
-    <link href="demo/app.css" rel="stylesheet">
+    <link href="home/home.css" rel="stylesheet">
     <base href="/">
   </head>
   <body>
