+++ demo/typings.d.ts
@@ -49,28 +49,3 @@ declare module 'angular2/testing' {
 declare module 'angular2/http/testing' {
   export * from '@angular/http/testing';
 }
-
-declare module tagInput {
-  interface TagInputComponent {
-    selectedTag: string;
-    model: {
-      value: string;
-      reset();
-    };
-    value: string[];
-    input: {
-      element: HTMLElement,
-      isFocused: boolean;
-      isVisible(): void;
-      focus(): void;
-    }
-    remove(item: string): void;
-    add(): void;
-    fireEvents(eventName: string, $event?: any);
-    handleKeydown($event, item: string);
-  }
-
-  interface TagComponent {
-    remove(): void;
-  }
-}
