+++ demo/vendor.ts
@@ -1,0 +1,8 @@
+// Vendors
+
+// Angular 2
+import '@angular/platform-browser-dynamic';
+import '@angular/platform-browser';
+import '@angular/core';
+import '@angular/http';
+import '@angular/router';
