+++ ng2-tag-input.ts
@@ -1,0 +1,9 @@
+import {TagInput}  from 'components/tag-input/tag-input.component';
+
+export {TagInput};
+
+export default {
+    components: [
+        TagInput
+    ]
+};
